+++ src/adapters/chashmap.rs
@@ -1,0 +1,51 @@
+use std::hash::Hash;
+use std::sync::Arc;
+
+use bustle::*;
+use chashmap::CHashMap;
+
+#[derive(Clone)]
+pub struct CHashMapTable<K>(Arc<CHashMap<K, u32>>);
+
+impl<K> Collection for CHashMapTable<K>
+where
+    K: Send + Sync + From<u64> + Copy + 'static + Hash + Eq + std::fmt::Debug,
+{
+    type Handle = Self;
+
+    fn with_capacity(capacity: usize) -> Self {
+        Self(Arc::new(CHashMap::with_capacity(capacity)))
+    }
+
+    fn pin(&self) -> Self::Handle {
+        self.clone()
+    }
+}
+
+impl<K> CollectionHandle for CHashMapTable<K>
+where
+    K: Send + Sync + From<u64> + Copy + 'static + Hash + Eq + std::fmt::Debug,
+{
+    type Key = K;
+
+    fn get(&mut self, key: &Self::Key) -> bool {
+        self.0.get(key).is_some()
+    }
+
+    fn insert(&mut self, key: &Self::Key) -> bool {
+        self.0.insert(*key, 0).is_none()
+    }
+
+    fn remove(&mut self, key: &Self::Key) -> bool {
+        self.0.remove(key).is_some()
+    }
+
+    fn update(&mut self, key: &Self::Key) -> bool {
+        self.0
+            .get_mut(key)
+            .map(|mut r| {
+                *r += 1;
+            })
+            .is_some()
+    }
+}
