+++ src/adapters/dashmap.rs
@@ -1,0 +1,51 @@
+use std::hash::{BuildHasher, Hash};
+use std::sync::Arc;
+
+use bustle::*;
+use dashmap::DashMap;
+
+#[derive(Clone)]
+pub struct DashMapTable<K, H>(Arc<DashMap<K, u32, H>>);
+
+impl<K, H> Collection for DashMapTable<K, H>
+where
+    K: Send + Sync + From<u64> + Copy + 'static + Hash + Eq + std::fmt::Debug,
+    H: BuildHasher + Default + Send + Sync + 'static + Clone,
+{
+    type Handle = Self;
+
+    fn with_capacity(capacity: usize) -> Self {
+        Self(Arc::new(DashMap::with_capacity_and_hasher(
+            capacity,
+            H::default(),
+        )))
+    }
+
+    fn pin(&self) -> Self::Handle {
+        self.clone()
+    }
+}
+
+impl<K, H> CollectionHandle for DashMapTable<K, H>
+where
+    K: Send + Sync + From<u64> + Copy + 'static + Hash + Eq + std::fmt::Debug,
+    H: BuildHasher + Default + Send + Sync + 'static + Clone,
+{
+    type Key = K;
+
+    fn get(&mut self, key: &Self::Key) -> bool {
+        self.0.extract(key, |_, _| ()).is_some()
+    }
+
+    fn insert(&mut self, key: &Self::Key) -> bool {
+        !self.0.insert(*key, 0)
+    }
+
+    fn remove(&mut self, key: &Self::Key) -> bool {
+        self.0.remove(key)
+    }
+
+    fn update(&mut self, key: &Self::Key) -> bool {
+        self.0.update(key, |_, v| v + 1)
+    }
+}
