+++ src/adapters/std.rs
@@ -1,0 +1,53 @@
+use std::collections::HashMap;
+use std::hash::{BuildHasher, Hash};
+use std::sync::Arc;
+
+use bustle::*;
+use parking_lot::RwLock;
+
+#[derive(Clone)]
+pub struct RwLockStdHashMapTable<K, H>(Arc<RwLock<HashMap<K, u32, H>>>);
+
+impl<K, H> Collection for RwLockStdHashMapTable<K, H>
+where
+    K: Send + Sync + From<u64> + Copy + 'static + Hash + Eq,
+    H: BuildHasher + Default + Send + Sync + 'static + Clone,
+{
+    type Handle = Self;
+
+    fn with_capacity(capacity: usize) -> Self {
+        Self(Arc::new(RwLock::new(HashMap::with_capacity_and_hasher(
+            capacity,
+            H::default(),
+        ))))
+    }
+
+    fn pin(&self) -> Self::Handle {
+        self.clone()
+    }
+}
+
+impl<K, H> CollectionHandle for RwLockStdHashMapTable<K, H>
+where
+    K: Send + Sync + From<u64> + Copy + 'static + Hash + Eq,
+    H: BuildHasher + Default + Send + Sync + 'static + Clone,
+{
+    type Key = K;
+
+    fn get(&mut self, key: &Self::Key) -> bool {
+        self.0.read().get(key).is_some()
+    }
+
+    fn insert(&mut self, key: &Self::Key) -> bool {
+        self.0.write().insert(*key, 0).is_none()
+    }
+
+    fn remove(&mut self, key: &Self::Key) -> bool {
+        self.0.write().remove(key).is_some()
+    }
+
+    fn update(&mut self, key: &Self::Key) -> bool {
+        let mut map = self.0.write();
+        map.get_mut(key).map(|v| *v += 1).is_some()
+    }
+}
